+++ duva/src/domains/cluster_actors/actor/tests/mod.rs
@@ -4,8 +4,7 @@ mod partitionings;
 mod replications;
 #[allow(unused_variables)]
 use super::actor::ClusterCommandHandler;
-use super::actor::cluster_actor_setups::FakeReadWrite;
-use super::session::SessionRequest;
+
 use super::*;
 use crate::CacheManager;
 use crate::NodeKind;
@@ -35,6 +34,64 @@ use tokio::net::TcpListener;
 use tokio::sync::mpsc::channel;
 use tokio::time::timeout;
 use uuid::Uuid;
+
+use std::sync::Arc;
+
+use crate::{
+    domains::{IoError, TRead, TWrite},
+    make_smart_pointer,
+};
+
+use bytes::BytesMut;
+use tokio::sync::Mutex;
+#[derive(Debug, Clone)]
+pub struct FakeReadWrite(Arc<Mutex<VecDeque<QueryIO>>>);
+make_smart_pointer!(FakeReadWrite, Arc<Mutex<VecDeque<QueryIO>>>);
+
+impl FakeReadWrite {
+    pub fn new() -> Self {
+        Self(Arc::new(Mutex::new(VecDeque::new())))
+    }
+}
+#[async_trait::async_trait]
+impl TWrite for FakeReadWrite {
+    async fn write(&mut self, io: QueryIO) -> Result<(), IoError> {
+        let mut guard = self.0.lock().await;
+        guard.push_back(io);
+        Ok(())
+    }
+}
+
+#[async_trait::async_trait]
+impl TRead for FakeReadWrite {
+    async fn read_bytes(&mut self, _buf: &mut BytesMut) -> Result<(), IoError> {
+        Ok(())
+    }
+
+    async fn read_values(&mut self) -> Result<Vec<QueryIO>, IoError> {
+        let mut values = Vec::new();
+        let mut guard = self.0.lock().await;
+
+        values.extend(guard.drain(..));
+        Ok(values)
+    }
+}
+
+pub(crate) fn create_peer_helper(
+    cluster_sender: ClusterCommandHandler,
+    hwm: u64,
+    repl_id: &ReplicationId,
+    port: u16,
+    node_kind: NodeKind,
+    fake_buf: FakeReadWrite,
+) -> (PeerIdentifier, Peer) {
+    let key = PeerIdentifier::new("127.0.0.1", port);
+
+    let kill_switch = PeerListener::spawn(fake_buf.clone(), cluster_sender, key.clone());
+    let peer =
+        Peer::new(fake_buf, PeerState::new(&key, hwm, repl_id.clone(), node_kind), kill_switch);
+    (key, peer)
+}
 
 fn write_operation_create_helper(
     index_num: u64,
@@ -121,6 +178,29 @@ fn consensus_request_create_helper(
     consensus_request
 }
 
+#[cfg(test)]
+impl<T: TWriteAheadLog> ClusterActor<T> {
+    pub(crate) fn test_add_peer(
+        &mut self,
+        port: u16,
+        kind: NodeKind,
+        repl_id: Option<ReplicationId>,
+    ) -> (FakeReadWrite, PeerIdentifier) {
+        let buf = FakeReadWrite::new();
+        let (id, peer) = create_peer_helper(
+            self.self_handler.clone(),
+            0,
+            &repl_id.unwrap_or_else(|| self.replication.replid.clone()),
+            port,
+            kind,
+            buf.clone(),
+        );
+
+        self.members.insert(id.clone(), peer);
+        (buf, id)
+    }
+}
+
 #[tokio::test]
 async fn test_requests_pending() {
     // GIVEN
@@ -130,7 +210,7 @@ async fn test_requests_pending() {
     let con_req = ConsensusRequest::new(write_r.clone(), tx, None);
 
     //WHEN
-    cluster_actor.test_block_write_reqs();
+    cluster_actor.block_write_reqs();
     cluster_actor.req_consensus(con_req).await;
 
     // THEN
@@ -140,4 +220,48 @@ async fn test_requests_pending() {
         cluster_actor.pending_requests.as_mut().unwrap().pop_front().unwrap().request,
         write_r
     );
+}
+
+#[tokio::test]
+async fn test_hop_count_when_one() {
+    // GIVEN
+    let fanout = 2;
+
+    // WHEN
+    let hop_count = ClusterActor::<MemoryOpLogs>::hop_count(fanout, 1);
+    // THEN
+    assert_eq!(hop_count, 0);
+}
+
+#[tokio::test]
+async fn test_hop_count_when_two() {
+    // GIVEN
+    let fanout = 2;
+
+    // WHEN
+    let hop_count = ClusterActor::<MemoryOpLogs>::hop_count(fanout, 2);
+    // THEN
+    assert_eq!(hop_count, 0);
+}
+
+#[tokio::test]
+async fn test_hop_count_when_three() {
+    // GIVEN
+    let fanout = 2;
+
+    // WHEN
+    let hop_count = ClusterActor::<MemoryOpLogs>::hop_count(fanout, 3);
+    // THEN
+    assert_eq!(hop_count, 1);
+}
+
+#[tokio::test]
+async fn test_hop_count_when_thirty() {
+    // GIVEN
+    let fanout = 2;
+
+    // WHEN
+    let hop_count = ClusterActor::<MemoryOpLogs>::hop_count(fanout, 30);
+    // THEN
+    assert_eq!(hop_count, 4);
 }
